Version 0.1.0
The mechanism works correctly: The user provides a list of products and a PDF is generated with stamps for the product.
The user has to execute the program from a terminal (like PowerShell).
The user must save the list of products as a txt file or docx file.
The user can choose which file to use as a data source to get the names of the products.
Finally, a subfolder is created in the results folder, which contains the PDF files with the stamps for each product.